+++ lib/widgets/post.dart
@@ -10,6 +10,7 @@ import 'package:url_launcher/url_launcher.dart' as ul;
 
 import '../hooks/delayed_loading.dart';
 import '../hooks/logged_in_action.dart';
+import '../hooks/stores.dart';
 import '../l10n/l10n.dart';
 import '../pages/create_post.dart';
 import '../pages/full_post.dart';
@@ -536,6 +537,8 @@ class _Voting extends HookWidget {
     final theme = Theme.of(context);
     final myVote = useState(post.myVote ?? VoteType.none);
     final loading = useDelayedLoading();
+    final showScores =
+        useConfigStoreSelect((configStore) => configStore.showScores);
     final loggedInAction = useLoggedInAction(post.instanceHost);
 
     vote(VoteType vote, Jwt token) async {
@@ -558,20 +561,21 @@ class _Voting extends HookWidget {
     return Row(
       children: [
         IconButton(
-            icon: Icon(
-              Icons.arrow_upward,
-              color: myVote.value == VoteType.up ? theme.accentColor : null,
+          icon: Icon(
+            Icons.arrow_upward,
+            color: myVote.value == VoteType.up ? theme.accentColor : null,
+          ),
+          onPressed: loggedInAction(
+            (token) => vote(
+              myVote.value == VoteType.up ? VoteType.none : VoteType.up,
+              token,
             ),
-            onPressed: loggedInAction(
-              (token) => vote(
-                myVote.value == VoteType.up ? VoteType.none : VoteType.up,
-                token,
-              ),
-            )),
+          ),
+        ),
         if (loading.loading)
           const SizedBox(
               width: 20, height: 20, child: CircularProgressIndicator())
-        else
+        else if (showScores)
           Text(NumberFormat.compact()
               .format(post.counts.score + (wasVoted ? 0 : myVote.value.value))),
         IconButton(